Fixed fps drops, fixed preset play, added error trappers, updated gitignore
There was a bug which had a huge impact on the app fps while running in the background, now it s been fixed. The play option for the preset was not working for the whole file, it was stopping the the first bookmark, now it works as intended. Added more try-catch statements in more places for ease in tracking runtime errors. Ignoring dev testing source codes (used to test c++ code such as function wrappers and more advanced c++ structures)